Move hook to hooks folder
