+++ packages/loaders/src/hooks/useCSSSVGAnimation.js
